Set module-name compiler option if able for api.make
Relative requires need a module name but `make` builds file-by-file
without `require` so no "true" module name can be known. Since we can
assume we're running in the context of neovim, where a `fnl/` folder is
good enough as a "package root", we can try to find a module name by
looking for that in the file path and processing as needed.

This may have false matches in some strange cases like
`/home/fnl/.../code/fnl/...` but for 99% of the time it will be fine and
only effects code that attempts to access `...` at compile time.