Replace vim.fn.* calls so Hotpot can run off the main thread
Viml may not be executed off the main thread. We only call it rarely for
a few helper functions and paths, so we now cache these values in the
runtime (which we can guess probably runs in the main thread at start
up) and do the rest (mkdir-p) with raw luv.