Do not search for lua modules when importing macros
This may be temporary or permanent. Because we now compile into the RTP,
its possible to fall into infinte recursion if a macro imports a module
with the same name (eg between "mod/init" and "mod/init-macros", both
resolving for "mod"). Making the macro searcher only find fnl files
avoids this as they are never technically in the RTP.