fix: expand longna~1 paths on windows for nvim_get_runtime_file
On some windows configurations, long usernames (or other dir names) are
truncated to "shortnames", and stdpath(cache) will return a path with
this truncation in it.

nvim_get_runtime_file cannot search these paths, which means if we add
this truncated path to the RTP, it will never find our compiled lua
inside the cache.

We can untruncate these paths by calling uv.fs_realpath, but that will
return nil if the path does not exist, so we must:

- first ensure that neovims cache dir exists ASAP, then
- when constructing our own paths inside the cache dir, call fs_realpath
  to untruncate the cache path.

https://github.com/neovim/neovim/issues/25019
https://github.com/rktjmp/hotpot.nvim/issues/118